# Implementing Regex optimization on the `MatchNotRegexp` matcher type