This PR brings the same regex optimization that already exists on the `MatchRegexp` to the `MatchNotRegexp` matcher.

With this change not regex matchers like `i!~"X|Y|Z` will be optimized to get the posting for `X`, `Y` and `Z` instead of evaluating the regex.


```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
                                                                     │   /tmp/old    │              /tmp/new               │
                                                                     │    sec/op     │    sec/op     vs base               │
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-32                         10.348m ± ∞ ¹   2.155m ± ∞ ¹  -79.17% (p=0.008 n=5)
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-32           10.544m ± ∞ ¹   2.169m ± ∞ ¹  -79.43% (p=0.008 n=5)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-32      130.6m ± ∞ ¹   132.3m ± ∞ ¹        ~ (p=0.222 n=5)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-32    204.0m ± ∞ ¹   204.9m ± ∞ ¹        ~ (p=0.310 n=5)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-32    671.1n ± ∞ ¹   678.0n ± ∞ ¹   +1.03% (p=0.008 n=5)
geomean                                                                 4.551m         2.437m        -46.45%
```